Add react as a peer dependency
If this package works with react 15 it would be very handy to add the react
version as a peer dependency to have the dependent supply the version of react.
In this case react 0.14 would be not installed in an app which is using react 15.
Currently it installs the react 0.14 additionally because it is not a peer dependency.